+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -64,7 +64,7 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-01-20T14:50:40.929Z")
+    created: Optional[datetime] = parse_datetime("2026-01-20T15:51:10.047Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")