+++ packages/orq-rc/src/orq_ai_sdk/models/reasoningpart.py
@@ -37,7 +37,7 @@ class ReasoningPart(BaseModel):
     r"""The reasoning or thought process behind the response. Used for chain-of-thought or extended thinking."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "reasoning_01kfdy64e97n3jvv3waexngh6n"
+        "reasoning_01kfe1mwg0n394524n5jxgv5hb"
     )
     r"""Unique identifier for the part. Format: reasoning_{ulid} (e.g., reasoning_01hxyz...)"""