+++ packages/orq-rc/src/orq_ai_sdk/models/retrievecontactop.py
@@ -103,7 +103,7 @@ class RetrieveContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-20T14:50:36.860Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-20T15:51:06.085Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")