+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFDY65K3TCNJ4XH9A3FPAYS3"
+        "01KFE1MXAJB0D2539Y8D8YKJCX"
     )
     r"""The unique identifier of the data source"""