+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFDY65KK3PK2BTZQK8BFR6BX"
+        "01KFE1MXAPJW257A3E3CN9QZ1G"
     )
     r"""The unique identifier of the data source"""